eatsmartly-backend/agents/brand_website_scraper.py
"""
Brand Website Scraper Agent
Attempts to fetch product information from official brand websites using barcode.
"""
import re
import asyncio
import requests
from typing import Dict, Any, Optional
from bs4 import BeautifulSoup
from agents.utils import setup_logger
from config import settings

logger = setup_logger(__name__, settings.LOG_LEVEL)


class BrandWebsiteScraper:
    """
    Scrapes official brand websites for product information.
    Uses barcode to search brand websites directly.
    """
    
    def __init__(self):
        """Initialize brand website scraper."""
        # Map of known brand patterns to their search URLs
        self.brand_search_urls = {
            # Indian Brands
            "amul": "https://www.amul.com/m/search?q=",
            "britannia": "https://www.britannia.co.in/products",
            "nestle": "https://www.nestle.in/brands",
            "itc": "https://www.itcportal.com/brands",
            "parle": "https://www.parleproducts.com/brand",
            "haldiram": "https://www.haldiram.com/products",
            
            # International Brands
            "coca-cola": "https://www.coca-cola.com/search?q=",
            "pepsi": "https://www.pepsico.com/search?q=",
            "unilever": "https://www.unilever.com/search/?q=",
            "nestle-global": "https://www.nestle.com/search?q=",
            "mondelez": "https://www.mondelezinternational.com/search?q=",
        }
        
        # Barcode prefix to brand mapping (GS1 company prefixes)
        self.barcode_to_brand = {
            "894": "Indian products",
            "890": "Indian products",
            "893": "Vietnamese products",
            "30": "US/Canada products",
            "50": "UK products",
            "76": "Swiss products (Nestle)",
            "40-43": "German products",
            "300-379": "French products",
            "80-83": "Italian products",
        }
        
        logger.info("BrandWebsiteScraper initialized")
    
    async def fetch_from_brand_website(self, barcode: str, product_name: Optional[str] = None) -> Optional[Dict[str, Any]]:
        """
        Try to fetch product info from official brand website.
        
        Args:
            barcode: Product barcode
            product_name: Optional product name to help identify brand
            
        Returns:
            Product data if found, None otherwise
        """
        try:
            logger.info(f"🌐 Attempting to scrape brand website for barcode: {barcode}")
            
            # Identify brand from barcode prefix
            brand_hint = self._identify_brand_from_barcode(barcode)
            logger.info(f"   Barcode prefix suggests: {brand_hint}")
            
            # If we have product name, try to extract brand
            if product_name:
                detected_brand = self._extract_brand_from_name(product_name)
                if detected_brand:
                    logger.info(f"   Detected brand from product name: {detected_brand}")
                    
                    # Try to scrape the brand website
                    result = await self._scrape_brand_site(detected_brand, barcode, product_name)
                    if result:
                        logger.info(f"✅ Successfully scraped {detected_brand} website")
                        return result
            
            logger.info(f"❌ Could not scrape brand website for barcode {barcode}")
            return None
            
        except Exception as e:
            logger.error(f"Error scraping brand website: {e}")
            return None
    
    def _identify_brand_from_barcode(self, barcode: str) -> str:
        """Identify likely country/brand from barcode prefix."""
        if len(barcode) < 3:
            return "Unknown"
        
        prefix = barcode[:3]
        
        # Check Indian prefixes
        if prefix in ["890", "893", "894"]:
            return "Indian brand (likely Amul, Britannia, ITC, or Parle)"
        
        # Check international prefixes
        prefix2 = barcode[:2]
        if prefix2 == "76":
            return "Swiss brand (likely Nestle)"
        elif prefix2 == "30":
            return "US/Canada brand"
        elif prefix2 == "50":
            return "UK brand"
        elif prefix2 in ["40", "41", "42", "43"]:
            return "German brand"
        
        return f"Unknown (prefix: {prefix})"
    
    def _extract_brand_from_name(self, product_name: str) -> Optional[str]:
        """Extract brand name from product name."""
        product_lower = product_name.lower()
        
        # Check known brands
        for brand in self.brand_search_urls.keys():
            if brand.replace("-", " ") in product_lower or brand in product_lower:
                return brand
        
        # Check for common Indian brands
        indian_brands = {
            "amul": "amul",
            "britannia": "britannia",
            "parle": "parle",
            "nestle": "nestle",
            "maggi": "nestle",
            "kitkat": "nestle",
            "nescafe": "nestle",
            "haldiram": "haldiram",
            "itc": "itc",
            "sunfeast": "itc",
            "aashirvaad": "itc",
        }
        
        for keyword, brand in indian_brands.items():
            if keyword in product_lower:
                return brand
        
        return None
    
    async def _scrape_brand_site(self, brand: str, barcode: str, product_name: str) -> Optional[Dict[str, Any]]:
        """
        Scrape specific brand website.
        
        This is a simplified implementation. In production, you'd need
        specific scrapers for each brand's website structure.
        """
        try:
            search_url = self.brand_search_urls.get(brand)
            if not search_url:
                logger.warning(f"No search URL configured for brand: {brand}")
                return None
            
            logger.info(f"   Searching {brand} website: {search_url}")
            
            # For now, return a placeholder
            # In production, implement actual scraping logic
            logger.warning(f"   Brand website scraping not yet implemented for: {brand}")
            logger.info(f"   Would search: {search_url}{product_name}")
            
            # Example structure of what we'd return:
            # return {
            #     "source": f"{brand.title()} Official Website",
            #     "barcode": barcode,
            #     "name": product_name,
            #     "url": f"{search_url}{barcode}",
            #     "scraped": True
            # }
            
            return None
            
        except Exception as e:
            logger.error(f"Error scraping {brand} website: {e}")
            return None
    
    def get_brand_website_url(self, barcode: str, product_name: Optional[str] = None) -> Optional[str]:
        """
        Get the URL to the product on the brand's official website.
        
        Args:
            barcode: Product barcode
            product_name: Optional product name
            
        Returns:
            URL string if brand is known, None otherwise
        """
        if not product_name:
            return None
        
        brand = self._extract_brand_from_name(product_name)
        if not brand:
            return None
        
        search_url = self.brand_search_urls.get(brand)
        if search_url:
            return f"{search_url}{barcode}"
        
        return None
